refactor: Move TransactionModal in its own folder
